// src/app/api/generated/models/tour-detail-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ComputedTourAttributesDto } from '../models/computed-tour-attributes-dto';
import { CoverImageDto } from '../models/cover-image-dto';
import { TourRouteDto } from '../models/tour-route-dto';
export interface TourDetailDto {
  computedAttributes?: ComputedTourAttributesDto;
  coverImage?: CoverImageDto;
  createdAt?: string;
  description?: string;
  endLocation?: string;
  estimatedDurationS?: number;
  id?: number;
  name?: string;
  plannedDistanceM?: number;
  route?: TourRouteDto;
  startLocation?: string;
  timezoneId?: string;
  transportType?: 'BIKE' | 'HIKE' | 'RUNNING' | 'VACATION';
  updatedAt?: string;
  userId?: number;
  version?: number;
}