// src/app/api/generated/models/update-tour-request.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { CoordinateDto } from '../models/coordinate-dto';
export interface UpdateTourRequest {
  description?: string;
  endCoordinate: CoordinateDto;
  endLocation: string;
  name: string;
  startCoordinate: CoordinateDto;
  startLocation: string;
  timezoneId: string;
  transportType: 'BIKE' | 'HIKE' | 'RUNNING' | 'VACATION';
  version: number;
}